Optionally enable sticky sessions
For some applications, subsequent requests for the same user or session
must be routed to the same container. This is because some libraries
like Socket.io for Node.js maintain session state within the container.
Although this is not a best practice, it exists in the wild and needs to
be accounted for.

This adds an option when generating the ingress stack or target group to
enable sticky sessions. When this is enabled, ALB will add a cookie to
requests so that it knows which target should be used for subsequent
requests.