Improve certificate management
Currently, the module does a good job of managing ACM certificates and
aliases for single account setups when all hosted zones are in the same
account. However, if hosted zones are spread out across accounts, you
need to disable certificate and alias management entirely. This reworks
the certificate management so it's possible to have a hybrid. It also
tries to be less clever when figuring out alternative domain names for
certificates, which should help prevent situations where it
unnecessarily detaches certificates from the load balancer during
changes.

- Clarify what's going to happen with each of the domain input variables
  by creating purpose-driven names like "issue_certificate_domains" and
  "create_domain_aliases."
- Make it possible to combine module-managed ACM certificates and
  aliases with externally created resources, so that the load balancer
  module itself can manage the primary SDLC domain and external
  certificates and aliases can be created separately.
- Provide better examples of multi-account and non-Route 53 setup.